Fix up the Makefile, add clean, install, check targets
The executable name can now by changed with TRGT, the SRCS array now
includes every file, and there is now an install target which supports
the standards PREFIX and DESTDIR variables commonly seen with
Autotools projects. The "make clean" target is capable of removing all
the objects.

Additionally, change the -std=c++0x to -std=c++11 which is the new
form that should be used instead.